+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog for Open Measurement SDK JavaScript clients
 
+## 1.6.3 - 2026-02-13
+- Add injection id attribute to injected script.
+- Make GUID generation a little more secure in environments where it's possible.
+
 ## 1.6.2 - 2026-01-06
 - Update chromedriver to 143.